requirements.txt
requests>=2.31.0
xgboost>=2.0.0
scikit-learn>=1.3.0
hmmlearn>=0.3.0
aiohttp>=3.9.0
websockets>=12.0
python-dotenv>=1.0.0
numpy>=1.24.0
scipy>=1.10.0
eth-account>=0.11.0
py-clob-client>=0.17.0
hexbytes>=0.3.0
